Allow setting null value through criteria API (#2475)
* Allow setting null value through criteria API

* Use CriteriaBuilder.literalNull for updating null values through criteria API

* Sonar fixes

+++ data-tck/src/main/groovy/io/micronaut/data/tck/tests/AbstractRepositorySpec.groovy
@@ -75,6 +75,8 @@ import static io.micronaut.data.tck.repositories.BookSpecifications.titleEqualsW
 import static io.micronaut.data.tck.repositories.PersonRepository.Specifications.distinct
 import static io.micronaut.data.tck.repositories.PersonRepository.Specifications.idsIn
 import static io.micronaut.data.tck.repositories.PersonRepository.Specifications.nameEquals
+import static io.micronaut.data.tck.repositories.PersonRepository.Specifications.setIncome
+import static io.micronaut.data.tck.repositories.PersonRepository.Specifications.setName
 
 abstract class AbstractRepositorySpec extends Specification {
 
@@ -2180,6 +2182,21 @@ abstract class AbstractRepositorySpec extends Specification {
             updated == 1
             personRepository.count(nameEquals("Xyz")) == 1
             personRepository.count(nameEquals("Jeff")) == 0
+        when:
+            personRepository.updateAll(setIncome(1000).where(nameEquals("James")))
+            def jamesPerson = personRepository.findByName("James")
+        then:
+            jamesPerson.income == 1000
+        when:"Income set to null using criteria"
+            personRepository.updateAll(setIncome(null).where(nameEquals("James")))
+            jamesPerson = personRepository.findByName("James")
+        then:"Field is updated to null"
+            jamesPerson.income == null
+        when:"Update name to null using criteria"
+            personRepository.updateAll(setName(null).where(nameEquals("James")))
+        then:"Exception is thrown because name is not nullable"
+            def ex = thrown(IllegalStateException)
+            ex.message == 'Field [name] does not allow null value.'
         when:
             deleted = personRepository.deleteAll(DeleteSpecification.where(nameEquals("Xyz")))
         then:
